Add Kubernetes demo module
This commit introduces a new kubernetes-demo module to the repository, providing practical and detailed example of how to work with core Kubernetes objects.

This new directory, located within the docker folder, contains manifest file (deployment.yaml) and a README.md to demonstrate how to declaratively deploy, expose, and manage applications. The addition of this module enriches the repository's resources, offering users a practical guide that complements the theoretical information on Kubernetes architecture and objects.